Fix unicode error in cert whitelist command
The string formatting here was not unicode-aware, so it would fail when
printing the whitelist for a course where one of the whitelisted users
had a unicode character in their username.

Django 1.8 does not officially support unicode usernames and will not
allow users to register as such. The bug was introduced when our support
staff manually fullfilled a user request to modify their name;
we should not do this.

Worse, since this code concatenated a giant string, instead of printing
out each line directly, this was more difficult to debug, as you
couldn't see which user was breaking things. The printing has been
updated output each user string individually.