Remove style guide from repository
This is now located in the wiki [1].

There were some reasons to keep this w/ the platform code;
it's convenient to always have access to both, etc.,
but it's a bit too much overhead to go through a platform pull request
just to update this. It's also more discoverable through the wiki, via
the initial landing page [2] and the sidebar. This means people are more
likely to see and hopefully update/extend it.

- [1] https://github.com/Stanford-Online/edx-platform/wiki/style-guide
- [2] https://github.com/Stanford-Online/edx-platform/wiki